+++ backend/src/main/java/ua/kvitkovo/orders/entity/OrderItem.java
@@ -1,11 +1,23 @@
 package ua.kvitkovo.orders.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
-import ua.kvitkovo.catalog.entity.Product;
-
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import java.math.BigDecimal;
 import java.util.Set;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import ua.kvitkovo.catalog.entity.Product;
 
 /**
  * @author Andriy Gaponov
@@ -39,7 +51,7 @@ public class OrderItem {
     private BigDecimal price;
 
     @Column(name = "product_qty")
-    private BigDecimal qty;
+    private int qty;
 
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "orderItem")
     private Set<OrderItemComposition> orderItemsCompositions;
